Changed/added the equals(Object o) methods for each class in the element package.
Added missing unsetXXX() methods.

Changed all the boolean, int, double, short variables in Boolean, Integer, Double and Short instances. However the interface is still the same.

All the getXXX() methods which should return a String returns an empty String if the variable to get is null.

All the getListOfXXX() can return null if the listOfXXX is not set.

It is still necessary to add toString() methods for each classes. Some stateChange() are still missing.