In many cases, the getModel() method was called twice, one time to check whether it leads to null and the second time to actually access methods of the model referenced by the current object. This is now changed to one call of getModel(), where we assign the result to a variable and then continue working with the reference of the model. The class Model also used to call getModel() in many places, where actually this.* was already sufficient. This has also been changed now.
The equals() method in Symbol was not correct and sometimes returned true where actually false was correct. The reason was that in cases when value was NaN, the variable equal was not considered anymore.

The SBMLWriter in package xml replaced the & symbol by the HTML code &amp;. This leads to problems if people want to include UTF8 characters in a note. We also recognized that HTML code sniplets cause problems when trying to write a model.

Bug fix: In cases of more than just one constructor, such as in AbstractSBase or SBMLDocument, we removed incorrect assignments by simply calling the default constructor first (this()).

The method get/setModelHistory() have been renamed to get/setHistory(). Only the Model still contains these older methods, which actually call the new methods from AbstractSBase.