Corrected a bug in listOf that sometimes caused errors due to invalid level/version combinations when adding elements.
Furthermore, the compartment class now overrides all setUnits and the setValue methods to ensure that an error is thrown if the spacial dimension is zero and an application tries to set the units or the value.

Furthermore, Compartment and Species now always call their initDefaults method in the constructor. This method actually initializes these objects and checks the level/version combination and sets all values as necessary.