Annotation has become a successor of StringTools for easier manipulation of its contained Strings.
In CVTerms only minor changes have been performed to better conform the Java coding style conventions. Furthermore, the Enumerations in CVTerm now have some methods that return values associated with it.

The code for the conversion to SI Units has been ported from libSBML 4.0.1. It has not yet been tested as this is work in progress.

BiologicalQuantifierParser was modified due to minor changes in CVTerm.