Some changes were necessary to avoid null pointer exceptions. Now, when accessing the annotation of an element for the first time, an empty annotation will be created instead of returning null. This is necessary because otherwise the try to get the list of CVTerms will result in a null pointer exception because CVTerms are gathered in the annotation. But the list of terms is allowed to be empty.
The UnitCompiler now returns invalid units as default and also the ASTNodeValue will have an invalid unit if requested. This is also necessary to avoid NullPointerExceptions. In this way, we can now in many cases already derive a unit, but often it is still not implemented.

Minor changes have been performed in UnitDefinition.