Now SBase extends the Swing interface TreeNode, which simplifies the creation of graphical interfaces to highlight the structure of SBML documents. Furthermore, the iteration over all subelements of each SBML element has now become possible.
Furthermore, I just noticed that in SBML Level 3 the exponent attribute has become a double number (before it was an integer). This is also important for the derivation of units of the SBML elements. The algorithm for this was improved.

Furthermore, the test package now includes a small GUI example that should read and display the content of a given SBML file to the user. However, it does currently not work in my test cases because the SBMLReader seems to add units to some elements before the corresponding UnitDefinition has been added to the model.