ASTNode : corrected the setType(String) to include almost everything. Missing only mathML annotations that are not used much from what I know. corrected also the compilation for the Type.LAMDBA to have the same output as in libSBML.
MathMLStaxParser : added support for the functionDefinition + few other elements like degree or logbase.

SBMLCoreParser, SBMLCoreElements.xml : added support for stochiometryMath.

ASTNode Compilers : added a method (lambdaFunction) to have the TextFormula parser print the functionDefinition as libSBML. Did not wanted to touch at the functions output so adding a new method was the easiest for me, we will have to see if we keep it or not.

MathML : corrected the lamdba function that was throwing and exception.