ASTNode : added two attribute to store the definitionURL and a boolean to know if the type of a number need to be written in mathML. Corrected the isBoolean method to return tru for relational operator as well. Corrected the type of power to FUNCTION_POWER.
MathMLStaxParser : reading the new ASTNode parameter now.

TextFormula : mainly added comments and TODO

MathMLXMLStreamCompiler : implemented almost all the mathML element (infinity is missing for example). The jsbml mathML output is now 99,99% the same as the libsbml mathML output. On the whole biomodels, there are only problems with the way that java read and transform the Double from a String. For example, 0.004 is becoming 0.0040, 1 is becoming 1.0 and more importantly 0.000166 is becoming 1.66e-4 which is illegal mathML/SBML if I wrote it back to a type="real" cn element.