SBase now indicates its property to be cloneable by extending the Java interface Cloneable. This doesn't change anything on our code but again improves the Java-like behavior of JSBML.
As announced previously, I introduced the new interface Assignment to unify all those elements that assign a value to be computed from an equation to some Variable. To this end, MathContainer had to become an interface and the additional class AbstractMathContainer now serves as the basic implementation for all elements containing ASTNode instances. In this way, an Assignment is also a MathContainer.

While performing this change, I noticed that the clone methods were not correct in all sub-classes of MathContainer and also that unsetVariable was missing in some Assignment types. This method has been added.

The web site where we usually created our class diagram was not available, so I transformed the code for class-diagram creation to graphviz code and created a new image that shows the relations of all SBML elements as it is implemented in JSBML now.