Renamed the documentation about differences between JSBML and LibSBML to a better title.
Exchanged the Stax2 and Woodstox libraries by newer versions.

Moved several special cases from the SBMLWriter where the name of the SBML element was created to the corresponding actual implementing class.

Fixed several bugs in many classes.

Changed the return type of the clone methods in some classes derived from AbstractSBase to the actual return type, i.e., the name of the class.

Corrected the way how indents were created when writing SBML.

Enabled that the new boolean attributes of Trigger can be written to SBML output.

Some classes, such as Priority, throw IllegalArgumentExceptions now if their usage is not permitted due to the SBML specification.