Notes : removed the notesBuffer with all the methods related to it, now only an XMLNode is store for the Notes.
readAttributes : corrected most of them (I did not checked them all) to return the correct boolean to say if the attribute was read or not. There is a warning message in SBMLCoreParser now to tell us if an attribute was not recognised.

removed the check for the level in AbstractMathContainer.setformula

SimpleSBaseChangeListener : replace the output to be done by log4j. You can control the output using the log4j.properties file.

JSBML : added an error message.

AbstracSBase : appendNotes(String), setNotes(XMLNode) and setNotes(String) are not yet fully implemented.

ASTNode : minor typos corrections

stax.SBMLWriter : changed the way Notes and Messages were written, to use the new XMLNode directly. Removed some TODO that were fixed/done.

MathMLStaxParser : removed the WritingParser interface methods that were not needed.

XMLNode : methods convertStringToXMLNode and convertStringToXMLNode not implemented.

Compartment : setVolume was calling setSize, so no possibility to set the volume in L1 !! I change it to call directly setValue()

Parameter : added a check for level in the initDefaults, before calling setConstant()

build.xml : corrected the bigjar task to include the classes as well as the sources