Deleted unnecessary SBO update script.
Moved log4j config file to resources folder.

Corrected problems in ASTNode when accessing nodes which are not defined.

Removed check method in ExplicitRule to unify the way how references are set amongst all instances of Assignment.

Added a TODO tag because the MathMLXMLStreamCompiler should also implement the ASTNodeCompiler interface for the sake of having identical standards.