SubModel, MathMLStaxParser : put back ASTNode.isName instead of ASTNode.isString
MathMLXMLStreamCompiler : added the namespace when writting XML elements.

ExplicitRule : added a bit of javadoc

ASTNode : remove some exception printing in the toString method.

History : cleaned some unused methods. Need to implement the toXML method as it is done for mathML.

SBMLWriter : corrected the methods writting the history.

DatesParser : corrected the way the created and modified dates were read as nothing was read before !!

SBMLCoreParser : change the class used to initialize the logger.

FunctionDefinition : small correction to the javadoc