Improved documentation, in particular the package descriptions within the src directory and also the document about the differences between JSBML and libSBML. This also includes the example code about how to create CellDesigner plug-ins easily.
Furthermore, I corrected the method to query for InitialAssignments by an id: it now searches for an initial assignment that has the given String as the value of its 'symbol' attribute and added a similar method for Rules as well.

To this end, I renamed the previous RuleVariableFilter to AssignmentVariableFilter and changed its documentation accordingly. This makes sense because we introduced the interface Assignment already.