Detected a bug in CVTerm: When Qualifiers where not set, i.e., null, the hashCode method created a NullPointerException. The two new methods to check whether the qualifiers are set properly avoid this.
I also improved the setNotes methods in case of given String elements: StringTools now contains the method that wraps a given String in the XML definition as required by SBML if necessary to be parsed later on. In this way, we can not only append (regular) Strings to existing XML nodes, but also create new XML comments from non-XML Strings.