Added the poster of the COMBINE in PDF and CorelDraw format.
Restructured the extension package: There is now one extra source folder for each extension. The core does no longer contain the ext package.
    
A first draft of the data structures for the spatial extension has been implemented.
  
When accessing the i-th child of an AbstractSBase, it is no longer necessary to sort all extension packages as these are now maintained in a sorted data structure.