The new interface TreeNodeWithChangeSupport separates manipulation of change listeners from SBase, allowing multiple implementations.
A check for duplicate identifiers has been implemented. JSBML now registers all elements that might have an identifier in hashes mapping the id to the object itself. Different rules depending on the specifications of the levels and versions of SBML ensure that only allowable identifiers can be accepted. To this end, the new interface UniqueNamedSBase tags all those elements whose id is intended to be unique within one model. A new JUnit test class has been implemented in order to check the capabilities of this id check.