When writing SBML Level 1 the name attribute (which has become the id attribute in later SBML versions) was never written because JSBML uses the id only as an internal representation. The check for Level 1 name attributes was not correct.
Increased version number of JSBML.

Solved runtime error that when extension sources were not included in the build path, JSBML couldn't be executed due to missing Parser classes.

Changed the data type of the map for the attributes to be written to TreeMap in order to ensure equal ordering of the attributes in every run. This is important for easy comparison of two generated XML files using diff.