TreeNodeChangeEvent : added a constant for a new method in SBase
AbstractSBase, SBase : added two news methods and one Map to be able to store XML namespace that were 
declared in XML or that should be declared on this element when writing the XML.

Model : update the getUnitDefinition(String) to be able to find the standard kinds.

SBMLCoreParser : updated the processNamespace method to be able to store namespace for 
	       all standard elements and Annotation element as well.

RDFAnnotationParser : update the processCharacter method to reflect the changes made in the way we parse annotations.

AnnotationParser : removed all the 'newline' that were added automatically in the code to replace them by
the real 'newline' made in the original XML files. Commented also the call to 'StringTools.encodeForHTML(characters)' as it was 
creating some strange characters.

SBMLReader : changed the way we read annotation. All the characters event are sent to the parser, the same way as we are doing for the notes. This allow to keep the same XML structure as in the original XML file. Also corrected a bug with some special annotation that were using RDF like the COPASI ones.

SBMLWriter : corrected the way we wrote annotations.

TODO.txt : clean and updated. could be removed ?