AbstractReaderWriter: corrected the way the listOf names where generated and now it will not adding a trailing 's' if the name of the element already finish in 's', like Species.
SBMLWriter : corrected the way we were writting annotations to always put it
in the SBML namespace, without prefix. There is still a general problem that we need to fix later,
the namespace of the children elements is the same by default as their parent which is not true
all the time, we should have one main namespace registered for every object and it should be this namespace
that we use when writting the file.

Qual classes : moved the namespaceURI and shortLabel of the package to the QualConstant class.

QualParser : adapted the class to use the new AbstractReaderWriter class.