MathMLXMLStreamCompiler : added some tests to check if we need or not to add the sbml namespace in case any cn elements defined some units.
FindUnitsCompiler : Added a small ASTNodeCompiler to check recursively for units in an ASTNode, most of the methods don't do anything.

JSBML : moved the method getNamespaceFrom from SBMLWriter to JSBML so that other classes can use it as well.

SBMLDocument : added a initDefaults() method to be able to set the SBML namespace.