extensions/layout/* : added a new parser for the layout L3 package. It seems to be working fine and I am able to read and write the layout_simple.xml file.
TreeNodeChangeEvent : moved some layout related constant to the LayoutConstant class.

Model : modified the default return value of Model.registerChild(SBase, SBase, boolean, boolean) to be able to register successfully object that are not NamedSBase when the recursive boolean is set to false (happen with listOf objects for example). For these, the AbstractSBase.registerChild(SBase) method was throwing an exception although there was no problem.

FindUnitsCompiler : removed some System.out