Bug fixes: ==========
Corrected problems of un-registering elements from SBMLDocument and Model (via id or metaid). Previously, it was possible that cloned elements or elements that were removed from the model at ealier states, could interfere with the register/unregister procedure causing that links to elements were deleted mistakenly.

Example: The element sbase was registered using its id in the model and using its metaid in the SBMLDocument.

In Model:        id     -> sbase
In SBMLDocument: metaid -> sbase

Now, a clone sbase' is created, sbase is removed from the model and the clone is registered. Note that now sbase' is part of the model and the pointers from the id will point to sbase'. However, sbase still has a pointer to its previous parent (which is important).

If we now remove any sub-element from sbase, pointers in Model and SBMLDocument from the ids/metaids to the sub-elements of sbase' will be deleted! Hence, a check was mounted to ensure that only correct elements will be unregistered.

The other changes are minor code/comment changes.


New features:
=============

The handling of user objects has been shifted to the interface TreeNodeWithChangeSupport and new methods were added for convenience.