Renamed LayoutConstant to LayoutConstants because there are multiple constants within this class.
Changed the default return type of the conversion method from Class<? extends SBase> to the listOf type from "none" to "other", because this is necessary for the packages. If some SBase is given, the listOf type should be "other", not "none". The default listOf type should still be "none", but not if a specific Class is given that is derived from SBase.