TestFunctionDefinition : added few tests to create and parse math infix formula or mathML containing a functionDefinition.
MathMLStaxParser : corrected the parser to be able to recognise a functionDefinition in mathML, even if the ASTNode is not linked to a model or if the functionDefinition id is wrong. We just check if there is an element 'ci' directly after an element 'apply'.

ASTNode : corrected a bug that created an infinite loop when using the FormulaParser to compile an incorrect ASTNode. If you have a call anywhere during the parsing of the ASTNode to the MathContainer.toString() method then this method call in turn the FormulaParser and create an infinite loop. There might be more of these call lying arround !!

I also changed the type return by the ASTNode.getParent method to be TreeNode instead of ASTNode, so that we can set the parentSBML object as the parent of top level ASTNode. Like this the tree structureof JSBML is working from an ASTNode as well and you can go up to the root, using the method getParent() only.