Corrected the CellDesignerAnnotationParser: No id has been set for the Layout element.
The SBMLcorrectorL3 was improved: More mandatory attributes are set and CellDesigner annotations are removed (because CellDesigner cannot open L3 Documents anyway).

Improved the way how yFiles was used to draw reactions: It is now possible to rotate the reaction nodes by arbitrary angles. However, this change might cause downstream effects in other programs and should therefore be double checked again. Furthermore, it seems that there is still a bug when computing these angles.

git-svn-id: svn://rarepos.cs.uni-tuebingen.de/SysBio/trunk@1107 e728d8a2-de87-4acf-9dc5-8dc47f4ff2f0