src/org/sbml/jsbml/AbstractSBase.java -- small typo correction
src/org/sbml/jsbml/SBMLDocument.java -- removed a namespace automatic addition, preparing to have only one namespace associated with each element.

src/org/sbml/jsbml/Model.java -- remove the specific getParent() method that was not compatible with the comp package. Does not seems to break any internal code but worth mentioning to the dev list may be and NEWS.

src/org/sbml/jsbml/xml/parsers/AbstractReaderWriter.java -- modified the writeNamespaces method to always set the correct namespace to the SBMLObjectForXML so that is it written correctly to XML. Also setting any declared namespaces.

PackageParserNamespaces.xml -- added the comp parser

SBMLCoreParser.java -- rewrote the getListOfSBMLElementsToWrite method to make use of the TreeNode interface to get the list of children of any elements easily. Also modified the writeNamespaces method to always set the correct namespace to the SBMLObjectForXML to avoid problem in complex cases with SBML packages (like comp).

src/org/sbml/jsbml/xml/stax/SBMLWriter.java -- major work done on the writeSBMLElements method, simplifying, cutting dead/unnecessary code. Now we use the WritingParser found with the namespace defined on the element we are writting instead to use the parent namespace to get the WritingParser. We need to make sure that all the package parser are working fine with this.