Large modifications of the layout code with 3 API changes that could break existing code:
- NamedSBaseGlyph was renamed to AbstractReferenceGlyph. The change seems logical to me when I made it
but it is not essential and could be reverted if people feel strongly about it.

- CurveSegment has been made Abstract and is now a superclass of LineSegment and CubicBezier. I have created a new enum to hold the two differents CurveSegment that we can have. When reading an XML file, we create only CurveSegmentImpl that correspond to the old CurveSegment with all children from LineSegment and CubicBezier defined. At the end of the reading, we convert those objects into LineSegment or CubicBezier instances as it should be.
Now when going through the listOfCurveSegments of a Curve, we would have to cast each element to either LineSegment or CubicBezier to access the specifics attributes/children. If people feel strongly about this change, we could introduce back all the getter and setter as abstract methods in CurveSegment and then throw an Exception if people use the methods [sg]etBasePoint[12] on LineSegment ?!.

- ExtendedLayoutModel, LayoutModelPlugin : inverted the inheritance to avoid having deprecated warning when using LayoutModelPlugin. This might break existing code in fact so if we let it like that we don't really need to keep the ExtendedLayoutModel class.

CompartmentGlyph : added the order attribute

CubicBezier : added a clone method from a CurveSegmentImpl + setting the type in constructors

Curve : did not removed yet the createCurveSegment() method that is creating a CurveSegmentImpl at the moment. Could be a problem if users don't set the type properly, although we could try to guess the type before writting. Added createLineSegment() and createCubicBezier() methods.

Dimensions : set the attributes to NaN in the constructors to have the methods isSetXXX() working properly + corrected the clone method to correctly set the attributes to NaN if they are not defined in the object to be cloned.


GraphicalObject : added the metaidRef attribute

Layout : using the class ListOfWithName for listOfAdditionalGraphicalObjects as it can contain different elements now.

LineSegment : added a clone method from a CurveSegmentImpl + setting the type in constructors

Point : corrected the clone method to correctly set the attributes to NaN if they are not defined in the object to be cloned.