Created a new ModelBuilder that facilitates the creation of SBML models by providing convenient methods that directly take multiple arguments to create model components.
A new class Pair has been added as a generalization of ValuePair. This class is not comparable and thus allows the creation of any 2-tuples of arbitrary objects. This is very useful, e.g., to create several SpeciesReferences by directly passing stoichiometry and species id to a build method. Furthermore, the static pairOf method can be imported into a calling class and hence greatly simplify the creation of parameterized data pairs.