Split out sections into separate files.
The early chapters followed a pattern of putting the \chapter and
lead-in text into User_Guide.tex, and putting sections into separate
files, but then later sections in the document didn't follow this
convention.  I found this confusing and annoying.  Either they should
all follow the pattern of the early chapters, or perhaps
alternatively, there should be only one file per chapter.  I opted for
the former approach, and split out sections into separate files.

Hopefully this is an improvement and not a step backward.