SBMLDocument unsets `fast` upon converting to L3V2 if possible.
The method `setLevelAndVersion` makes now use of a
`LevelVersionConverter`, which is an abstract class that for now has
just an implementation for upconverting to level 3 and only limited
functionality. When level and version of an `SBMLDocument` are set to
(3, 2), JSBML recursively traverses the full data structure. Before this
current change, it only tried to update level/value pairs in every model
component. Now it checks if an action needs to be performed on every
model component. If this is the case, e.g., for `Reaction` it checks
what needs to be changed. If a reaction has `fast=true` the method
`setLevelAndVersion` will return `false` because there is no equivalent
in L3V2. If fast is set to `false` it will silently remove the attribute
and return `true`.