Continued work on collecting into single SBMLDocument
Encountered difficulty: Resolving the SBMLDocument wherein a ModelDefinition/Submodel is found is currently not possible through getSBMLDocument -- i am currently using a hacky placeholder-approach via UserObjects, that will need be replaced. I think, when reading an SBMLDocument, the corresponding field to getSBMLDocument (in the root SBase) and the parent-child-structure of the nodes should be set, which it seems not to be currently for comp-files.
I have added a very simple test-case for recursive externalModelDefinitions where the internaliseExternalModelDefinitions-method has to show some interesting behaviour: it currently fails (bc. implementation is todo); more complicated test-cases are still to do.