Removed absolute...-methods
! Will discuss with Dräger.
Reasoning: The change from D/double to RelAbsVector already changes the API in such a way, that code using the previous API will not work anymore and will need be fixed (with the benefit, that Code for libSBML will be easier to switch to jsbml). Since the code will need be fixed anyways, one might as well already remove the now obsolete (and also non-specification-conform) absolute...-methods/fields etc, so that code using them can be updated appropriately ONCE, instead of deprecating the latter, and removing it later. Backwards-compatibility is not a possibility due to Double vs RelAbsVector.

Also, the new implementation should be able to deal with actual relabsvectors in SBMLfiles, where the old would have had difficulty/errors