Fix for issue #209: broken logo images
This points to images on github, which is not the best solution, but
at least it's more stable than the sbml.org pagespeed cache images
referenced before.  We should put a copy of the logo in the apidocs
output directory somewhere, but I couldn't immediately figure out how
to do that, and this was faster, so here it is.